+++ GiveAndTake/Assets/scripts/PuzzleVerification.cs
@@ -115,7 +115,7 @@ public class PuzzleVerification : MonoBehaviour
                 if (Star3.activeSelf)
                 {            
                     //buscar o PointsEarned e colocar o valor do scoreManager.GetScore() nele, junto com uma frase
-                    PointsEarned.text = "Conseguiste ganhar um total de " + scoreManager.GetScore() + " pontos!";
+                    PointsEarned.text = "Conseguiste ganhar um total de " + (int) scoreManager.GetScore() + " pontos!";
                     darkerBackground.SetActive(true);
                     PopUpStars.SetActive(true);
                     PopUpTechinic.SetActive(true);
@@ -124,7 +124,7 @@ public class PuzzleVerification : MonoBehaviour
                 }
                 else if (Star2.activeSelf)
                 {
-                    PointsEarned.text = "Conseguiste ganhar um total de " + scoreManager.GetScore() + " pontos!";
+                    PointsEarned.text = "Conseguiste ganhar um total de " + (int) scoreManager.GetScore() + " pontos!";
                     darkerBackground.SetActive(true);
                     PopUpStars.SetActive(true);
                     PopUpTechinic.SetActive(true);
@@ -133,7 +133,7 @@ public class PuzzleVerification : MonoBehaviour
                 }
                 else if (Star.activeSelf)
                 {
-                    PointsEarned.text = "Conseguiste ganhar um total de " + scoreManager.GetScore() + " pontos!";
+                    PointsEarned.text = "Conseguiste ganhar um total de " + (int) scoreManager.GetScore() + " pontos!";
                     darkerBackground.SetActive(true);
                     PopUpStars.SetActive(true);
                     PopUpTechinic.SetActive(true);
